# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  model.BackupLog:
    properties:
      completedAt:
        type: string
      createdBy:
        type: string
      fileName:
        type: string
      filePath:
        type: string
      fileSize:
        type: integer
      id:
        type: string
      notes:
        type: string
      startedAt:
        type: string
      status:
        type: string
    type: object
...
  response.PaginatedData:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.CreateBackupRequest:
    properties:
      created_by:
        type: string
      notes:
        type: string
    type: object
...
host: localhost:20800
info:
  contact:
    email: support@vnet.net
    name: VNET Support
  description: Hệ thống Quản lý Phòng Game Toàn Diện
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://vnet.net/terms
  title: VNET Core API
  version: "1.0"
paths:
...
  /api/backups:
    get:
      consumes:
      - application/json
      description: Get a paginated list of database backups
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/response.PaginatedData'
                  - properties:
                      data:
                        items:
                          $ref: '#/definitions/model.BackupLog'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List backups
      tags:
      - Backups
    post:
      consumes:
      - application/json
      description: Create a new database backup
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CreateBackupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.BackupLog'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create backup
      tags:
      - Backups
  /api/backups/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a database from a backup
      parameters:
      - description: Backup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Restore backup
      tags:
      - Backups
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"